+++ packages/plugin-hardhat/src/deploy-proxy.ts
@@ -51,6 +51,7 @@ export function makeDeployProxy(hre: HardhatRuntimeEnvironment, defenderModule: 
 
     const contractInterface = ImplFactory.interface;
     const data = getInitializerData(contractInterface, args, opts.initializer);
+    const deployFn = opts.deployFunction || deploy;
 
     if (await manifest.getAdmin()) {
       if (kind === 'uups') {
@@ -79,8 +80,8 @@ export function makeDeployProxy(hre: HardhatRuntimeEnvironment, defenderModule: 
           throw new InitialOwnerUnsupportedKindError(kind);
         }
 
-        const ProxyFactory = await getProxyFactory(hre, signer);
-        proxyDeployment = Object.assign({ kind }, await deploy(hre, opts, ProxyFactory, impl, data));
+        const ProxyFactory = opts.proxyFactory || (await getProxyFactory(hre, signer));
+        proxyDeployment = Object.assign({ kind }, await deployFn(hre, opts, ProxyFactory, impl, data));
         break;
       }
 
@@ -95,10 +96,11 @@ export function makeDeployProxy(hre: HardhatRuntimeEnvironment, defenderModule: 
           );
         }
 
-        const TransparentUpgradeableProxyFactory = await getTransparentUpgradeableProxyFactory(hre, signer);
+        const TransparentUpgradeableProxyFactory =
+          opts.proxyFactory || (await getTransparentUpgradeableProxyFactory(hre, signer));
         proxyDeployment = Object.assign(
           { kind },
-          await deploy(hre, opts, TransparentUpgradeableProxyFactory, impl, initialOwner, data),
+          await deployFn(hre, opts, TransparentUpgradeableProxyFactory, impl, initialOwner, data),
         );
         break;
       }